correção dos testes que fazem persistencia de entidades que contém municipios; inclusão do campo de depósito; alteração do tipo do campo "cadência" no processo de cotação de frete

+++ BsBios.Portal.ViewModel/OrdemDeTransporteCadastroVm.cs
@@ -30,7 +30,7 @@ namespace BsBios.Portal.ViewModel
         public string Classificacao { get; set; }
 
         [DisplayName("Cadência: ")]
-        public int Cadencia { get; set; }
+        public decimal Cadencia { get; set; }
         [DisplayName("Requisitos: ")]
         public string Requisitos { get; set; }
